Move copyURL into a separate utility file
